requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorgram.py==1.2.0
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
phue==1.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
spotipy==2.13.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
